% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quack_a_geddon.R
\name{quack_a_geddon}
\alias{quack_a_geddon}
\title{Escape the emotional Quack-A-Geddon}
\usage{
quack_a_geddon()
}
\description{
Escape the debugging induced emotional Quack-A-Geddon and let
the wise duck of the magic pond fill you with the motivation to continue
your quest for bug free code.

Once your inner pond has settled and your drive returns you should check out
\code{\link{quack_debug}} to gain a companion on your quest towards bug free
code.
}
\examples{
quack_a_geddon()
}
