% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quack_debug.R
\name{quack_debug}
\alias{quack_debug}
\title{Rubber duck debugging}
\usage{
quack_debug()
}
\description{
Use the power of rubber duck debugging to find bugs faster. Simply
talk the duck through your code by explaining the logic line by line and be
surprised by the EURICA moment once you find the bug by talking to a virtual
duck.

During the conversation you can type "done" to exit.

If the bug is persistent and your patience is running low, take a break and
enjoy some emotional support by running \code{\link{quack_a_geddon}}
}
\examples{
quack_debug()
}
