% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackR-package.R
\docType{package}
\name{quackR}
\alias{quackR-package}
\alias{quackR}
\title{quackR: Emotional Support and Rubber Duck Debugging}
\description{
The 'quackR' package provides a suite of interactive tools to help
developers through the psychological hurdles of debugging. By utilizing the
'Wise Duck' and the principles of
\href{https://en.wikipedia.org/wiki/Rubber_duck_debugging}{Rubber Duck Debugging},
users can verbalize their logic, find bugs and regain their will to code.
}
\section{Key Functions}{

\itemize{
  \item \code{\link{quack_debug}}: Start an interactive debugging session.
  \item \code{\link{quack_a_geddon}}: Summon the Wise Duck for emotional rescue.
}
}

\author{
\strong{Maintainer}: Julian Präsent \email{julian.praesent@gmail.com}

}
